ev3/requirements.txt
psutil==5.8.0
pythonping==1.1.0
paho-mqtt==1.5.1
watchdog==2.1.5
timeloop==1.0.2
requests==2.26.0
logging==0.4.9.6
pyyaml==5.4.1